Support multiple cards & arrays, and auto-detect.
Previously this plugin would only look at card 0, unit 0, which means it was of
little use if you have multiple RAID controllers, or multiple arrays
configured.  Also, the card number assigned seems to be somewhat arbitrary, and
I've been told it can sometimes change when the machine is rebooted, too.

Refactored to fetch a list of cards present using tw_cli, and for each card,
fetch the status of the arrays on that card and check they're all OK.

Also added an option to be verbose, to make sure that it's functioning
correctly, for instance:

[dave@clooney:~]$ perl nagios_3ware_raid_check --verbos
Card 4 is a 8006-2LP
Unit u0 on card4 is a RAID-1 array of 931 GB and is OK
RAIDCHECK OK - Card 4 unit u0 RAID-1 OK (931GB)

Finally, added the ability to pass the path to the tw_cli binary using the new
`--tw_cli` (or `--tw-cli`, or `-t`) option, if tw_cli isn't in your path.