Adding Bot::BasicBot::Pluggable module SayFromQueue.
A simple module to watch a directory for new files appearing and announce their
contents on IRC.

Allows other local processes to have the bot say things on IRC on their behalf,
by dropping a file containg e.g. "#channel:Hello there!" with a random name into
the directory the bot watches.

Could do with refactoring to allow the message queue dir to be configured.

I might also write a replacement that listens on a TCP or Unix socket for
messages.