Change setting name to github_project; docs fix.
I'm rocking up another module that is going to share the github_project setting
with this module.

In time, I may refactor the code to determine which module is default out to
share it between the two, and support different default projects per-channel, so
the same bot could default to different projects depending which channel it's
being asked in.