Fetch modem event log entries too.
Fetch event log entries from modem, store them in the DB, then clear the modem's
event log so that we don't keep seeing the same entries multiple times.

This is a bit of a dirty approach; there's a potential race condition for events
appearing between us requesting the event log page and clearing the event log -
those events would be lost.

I don't think it's a big deal, but it's not perfect.

Includes a refactor to use LWP::UserAgent instead of LWP::Simple, as I need to
make a POST request, which LWP::Simple annoyingly doesn't do.  As a benefit,
though, the authentication details can just be set up with a call to
$ua->credentials once, rather than passed in every URL, so that's a little bit
of win to brighten my evening (wow, what an exciting evening so far, eh? :) )