Nagios plugin to monitor SMART error log.
I was using check_ide_smart from Nagios.

It reported all good on a drive that was clearly failing and throwing errors,
because SMART's overall health asssessment for the drive was still a pass.

That's fucking useless.

    [dave@devvps:~]$ sudo /usr/lib/nagios/plugins/check_ide_smart -n /dev/sdb
    OK - Operational (18/18 tests passed)

This plugin monitors the contents of the SMART error logs, rather than relying
on SMART's asssment:

    [dave@devvps:~]$ ./nagios_smart_log -d /dev/sdb
    CHECKSMARTLOG CRITICAL - 5 SMART errors, latest: Error 13 occurred at disk
    power-on lifetime: 29769 hours (1240 days + 9 hours)

In other words, it produces information which is actually useful to identify a
failing drive, rather than listening to SMART happily proclaiming all is well
when it absolutely so fucking clearly isn't.