Detect if backup is currently running.
Check if a backup is currently running; if so, report OK.

TODO: check how long the backup has been running for, and alarm if it's been too
long.

Alternatively, skip using the rdiff-backup command to check for increments, and
simply dig around in the `rdiff-backup-data` dir for increments and work it out
ourselves, so we know when the last successful backup was completed even if a
backup is currently running.