New log-vm-superhub-stats.
Based on log-ntl-cablemodem-stats as already found in this repo at:
https://github.com/bigpresh/misc-scripts/tree/master/log-ntl-cablemodem-stats

I've just received a Virgin Media "SuperHub" (Netgear-powered bit of kit) as
part of a broadband speed upgrade - this is an updated version of the above
script, to fetch signal statistics from the new SuperHub.

It's a bit rough, to be fair; it's late at night and I've been in the pub.  It
works, though.

It records the useful info into the DB with the same key names as the old
version, so historical data analysis will still work for me.

This is primarily written for my own use, but released here in case it's of use
to anyone else.  If you do find it of use, feel free to let me know - it may
encourage me to polish it :)