Apply auth patch from Jeff Fern.
Jeff contacted me regarding this script, and was finding that authentication to
the modem wasn't working.  After some troubleshooting, he arrived at this
solution, which worked for him.  I haven't been able to test it as I no longer
use my NTL/Ambit 120 cablemodem.  So, I'm committing it here; if you want to use
this script, and find that you have authentication problems, you may wish to try
checking out the script as it was before this commit, and see if that works for
you.

(Please feel free to let me know, too!)