Adding photo-frame-populate script.
A script to walk through an input directory, and for each file/symlink therein,
check if we already have a file of that name in our output directory, and if
not, invoke ImageMagick's convert utility to resize the source image down,
auto-rotate it if necessary, and write it to the destination.

I use this for my photo frame (a 12-inch NIX X12C unit) which doesn't appear to
rotate the photos by the EXIF info, and doesn't require the full resolution of
my camera - it sure as hell doesn't have that high a resolution!