A munin plugin to graph stats from LagMeter
The Minecraft LagMeter server plugin can write a log file of lag-related details
at regular intervals, useful to look at what was going on at certain times when
lag was reported.

This is a quick and dirty Munin plugin to read the most recent entry from that
log and return the requested value to be graphed (number of players online,
chunks loaded, total entities, or TPS supported in this initial version).