Support max_price in plane monitoring specs.
Useful to filter to only prices you'd consider good, to keep an eye out for
bargains.

I may extend it in future to record each new plane for sale to a DB, so we could
have more clever things, e.g. comparing the price to the average market price
we've seen, or maybe pulling the base price per model and comparing against
that, too - but this'll be useful for now, and is easy!