Quick dirty script to watch a Sainsburys product
Need to buy more school polos for Abbie; Sainsburys was one of not many
places we found that did nice girls polos small enough for Abbie, but we
could only get two packs and they're out of stock on their website.

They don't provide any way to sign up to be alerted if it comes into
stock, so a quick rough & ready script to do it myself.

This should really be better written, controlled by command line options
etc rather than hardcoded, but it'll do for noe.