Get rtl_433 parsing working again.
Update which temperature sensor to watch, also receive transmissions
from my Watchman oil level monitor.

Still need to figure out what the values from the Watchman represent, and how
to turn them into a % of tank capacity left.  The depth_cm looks
promising, but over a few days of data, it's varied only a tiny bit,
going *up* as well as down (unless that's just the effect of the
temperature on the oil, i.e. expanding on hotter days?)