Fix "already seen" handling, remember when we saw it
D'oh, logic fail - of course, `$fetch_sth->execute` will always be true
(unless executing the query failed) - that doesn't mean we actually found
anything.

Check for an existing record properly - and, while I'm changing it,
record when we first see a record, and output that in our "already saw"
message.