Add ability to run rdiff-backup with sudo
Add the ability to run rdiff-backup as root to look up most recent
incrememental backup, useful if backups aren't accessible as the user
Nagios will run as.

May extend further such that it allows you to tell it to run as a given
user perhaps, but for now this works well.